Moved database code to 'services'
